+++ run.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
   
     _query_params = {
         "schema": "public",
-        "table_name": "film_list",
+        "table": "film_list",
         "columns": ["category", "price"],
         "aggregate": {
             "price": "SUM"
@@ -35,7 +35,12 @@ if __name__ == "__main__":
     # Using SimplePgSQL class
     pgsql = SimplePgSQL(conn_params, return_type=pd.DataFrame)
 
-    q_results = pgsql.execute("SELECT category, price FROM film_list LIMIT 10;", columns=["category", "price"])
+    _query = """
+            SELECT category, SUM(price) from public.film_list where length > 60 group by category, price order by price DESC limit 10;
+            """
+
+
+    q_results = pgsql.execute(_query, columns=["category", "price"])
     r_results = pgsql.read(**_query_params)
     print(q_results)
     print(r_results)
@@ -43,4 +48,8 @@ if __name__ == "__main__":
     # Deprecated method to query data DO NOT USE. For backward compatibility only
     with DBConnect(conn_params, return_type=pd.DataFrame) as cursor:
         q_results = cursor.query("SELECT category, price FROM film_list LIMIT 10;", columns=["category", "price"])
+        _query_params.pop("table")
+        _query_params["table_name"] = "film_list"
         r_results = cursor.read(**_query_params)
+        print(r_results)
+        print(q_results)