+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.1.8] - 2024-04-24
+
+### Bugfix
+1. Kept Naming Scheme consistent.
+2. Fixed typo for conjuction > conjunction
+
 ## [0.1.7] - 2024-04-16
 
 ### Added